refactor: seperate api method by route level
